requirements.txt
anyio==4.13.0
certifi==2026.2.25
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
psycopg2-binary==2.9.11
python-telegram-bot==20.7
sniffio==1.3.1
typing_extensions==4.15.0
